Add files via upload
This research focuses on the potential threats to
a website, such as DDOS attacks, SQL injection, and buffer
overflow. In this project, we created a complete website with
a database and tested some common risks. We used the Splunk
platform to do data analysis on the data generated by the attacks
in order to quickly identify attacks and predict possible threats.
The results of the analysis were utilized to determine how to
respond to these attacks and propose solutions.